Refactor CLI stats and corpus commands
